+++ include/net/af_rxrpc.h
@@ -30,13 +30,13 @@ enum rxrpc_interruptibility {
 extern atomic_t rxrpc_debug_id;
 
 typedef void (*rxrpc_notify_rx_t)(struct sock *, struct rxrpc_call *,
-				  unsigned long);
+				  uintptr_t);
 typedef void (*rxrpc_notify_end_tx_t)(struct sock *, struct rxrpc_call *,
-				      unsigned long);
+				      uintptr_t);
 typedef void (*rxrpc_notify_new_call_t)(struct sock *, struct rxrpc_call *,
-					unsigned long);
-typedef void (*rxrpc_discard_new_call_t)(struct rxrpc_call *, unsigned long);
-typedef void (*rxrpc_user_attach_call_t)(struct rxrpc_call *, unsigned long);
+					uintptr_t);
+typedef void (*rxrpc_discard_new_call_t)(struct rxrpc_call *, uintptr_t);
+typedef void (*rxrpc_user_attach_call_t)(struct rxrpc_call *, uintptr_t);
 
 void rxrpc_kernel_new_call_notification(struct socket *,
 					rxrpc_notify_new_call_t,
@@ -44,7 +44,7 @@ void rxrpc_kernel_new_call_notification(struct socket *,
 struct rxrpc_call *rxrpc_kernel_begin_call(struct socket *sock,
 					   struct rxrpc_peer *peer,
 					   struct key *key,
-					   unsigned long user_call_ID,
+					   uintptr_t user_call_ID,
 					   s64 tx_total_len,
 					   u32 hard_timeout,
 					   gfp_t gfp,
@@ -71,7 +71,7 @@ const struct sockaddr_rxrpc *rxrpc_kernel_remote_srx(const struct rxrpc_peer *pe
 const struct sockaddr *rxrpc_kernel_remote_addr(const struct rxrpc_peer *peer);
 unsigned int rxrpc_kernel_get_srtt(const struct rxrpc_peer *);
 int rxrpc_kernel_charge_accept(struct socket *, rxrpc_notify_rx_t,
-			       rxrpc_user_attach_call_t, unsigned long, gfp_t,
+			       rxrpc_user_attach_call_t, uintptr_t, gfp_t,
 			       unsigned int);
 void rxrpc_kernel_set_tx_length(struct socket *, struct rxrpc_call *, s64);
 bool rxrpc_kernel_check_life(const struct socket *, const struct rxrpc_call *);
